Adding tests for `time_zone` parameter for date range aggregation
